+++ Data/WrathClassic/ExportedCategories.lua
@@ -56,7 +56,7 @@ local a = addon;
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2024-03-12 19-17-09 ]] --
+-- [[ Exported at 2024-03-20 19-45-42 ]] --
 tasks = {
     {F, 1557, "UncategorizedCategories"},
     {C, 971, 1557},
@@ -98,6 +98,8 @@ tasks = {
     {A, 1346, 3478}, -- Pilgrim
     {C, 971, 1346},
     {N, 1346, a.L["Ta's Pet Collection"]}, -- Ta's Pet Collection
+    {C, 1566, 1572},
+    {N, 1572, a.L["Season"] .. " " .. 4, true}, -- Season 4
     {C, 1566, 1569},
     {N, 1569, a.L["Season"] .. " " .. 3, true}, -- Season 3
     {C, 1566, 1568},