+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 69.1 - 2024-03-20
+### Fixed
+- Localization
+
+## 69.0 - 2024-03-20
+### Added
+- New Back from the Beyond achievement
+- Plunderstorm achievements (2)
+- Noblegarden achievements (1)
+- Dragonflight Mythic+ Season 4 achievements (13)
+- Dragonflight PvP Season 4 achievements (18)
+- Awakened Raids achievements (12)
+- Misc achievements (9)
+- Dragon racing achievements (1)
+
+### Changed
+- Back from the Beyond is now a legacy achievement
+
 ## 68.4 - 2024-03-12
 ### Fixed
 - Moved Just an Ordinary Gas Cloud from Dragonflight - Profession - Engineering to The Burning Crusade - Profession - Engineering